refactor(extension-common): リンク検証を移す
広告リンクの OPID 照合、検証状態のストア、警告ページへのリダイレクトを
extension-common へ移す。

警告ページの場所はアプリごとに異なるため、URL の組み立てを
createLinkVerificationHandlers() の引数で受け取る形に変えた。従来は
chrome.runtime.getURL("index.html") と #/warning を直接組み立てていた。
