+++ packages/extension-common/src/env.d.ts
@@ -1,0 +1,6 @@
+interface ImportMeta {
+  env: {
+    /** ビルドモード。アプリのビルド時に注入される */
+    MODE: "development" | "production" | "testing";
+  };
+}
